+++ wxc_sdk/cdr/__init__.py
@@ -103,7 +103,6 @@ class CDRUserType(str, Enum):
 
 class CDR(ApiModel):
 
-    # TODO: CDR API returns empty strings instead of null for unset values. Makes it hard to consume
     @root_validator(pre=True)
     def force_none(cls, values: dict):
         """
@@ -121,41 +120,35 @@ class CDR(ApiModel):
     #: Whether the call leg was answered. For example, in a hunt group case, some legs will be unanswered,
     # and one will be answered.
     answered: Optional[bool] = Field(alias='Answered')
+    #: Whether the call was inbound or outbound. The possible values are:
+    direction: Optional[Union[CDRDirection, str]] = Field(alias='Direction')
+    #: For incoming calls, the calling line ID of the user. For outgoing calls, it's the calling line ID of the
+    #: called party.
+    called_line_id: Optional[str] = Field(alias='Called line ID')
     #: SIP Call ID used to identify the call. You can share the Call ID with Cisco TAC to help them pinpoint a call
     # if necessary.
     call_id: Optional[str] = Field(alias='Call ID')
-    #: Type of call. For example:
-    call_type: Optional[Union[CDRCallType, str]] = Field(alias='Call type')
-    #: For incoming calls, the calling line ID of the user. For outgoing calls, it's the calling line ID of the
-    # called party.
-    called_line_id: Optional[str] = Field(alias='Called line ID')
-    #: For incoming calls, the telephone number of the user. For outgoing calls, it's the telephone number of the
-    # called party.
-    called_number: Optional[str] = Field(alias='Called number')
     #: For incoming calls, the calling line ID of the calling party. For outgoing calls, it's the calling line ID of
     # the user.
     calling_line_id: Optional[str] = Field(alias='Calling line ID')
-    #: For incoming calls, the telephone number of the calling party. For outgoing calls, it's the telephone number
-    # of the user.
-    calling_number: Optional[str] = Field(alias='Calling number')
+    #: This is the start time of the call, the answer time may be slightly after this. Time is in UTC.
+    start_time: Optional[datetime] = Field(alias='Start time')
+    #: Type of call. For example:
+    call_type: Optional[Union[CDRCallType, str]] = Field(alias='Call type')
     #: The type of client that the user (creating this record) is using to make or receive the call. For example:
     client_type: Optional[Union[CDRClientType, str]] = Field(alias='Client type')
     #: The version of the client that the user (creating this record) is using to make or receive the call.
     client_version: Optional[str] = Field(alias='Client version')
     #: Correlation ID to tie together multiple call legs of the same call session.
     correlation_id: Optional[str] = Field(alias='Correlation ID')
+    #: The country code of the dialed number. This is only populated for international calls.
+    international_country: Optional[str] = Field(alias='International country')
     #: The MAC address of the device, if known.
     device_mac: Optional[str] = Field(alias='Device MAC')
-    #: Whether the call was inbound or outbound. The possible values are:
-    direction: Optional[Union[CDRDirection, str]] = Field(alias='Direction')
     #: The length of the call in seconds.
     duration: Optional[int] = Field(alias='Duration')
     #: Inbound trunk may be presented in Originating and Terminating records.
     inbound_trunk: Optional[str] = Field(alias='Inbound trunk')
-    #: The country code of the dialed number. This is only populated for international calls.
-    international_country: Optional[str] = Field(alias='International country')
-    #: Location of the report.
-    location: Optional[str] = Field(alias='Location')
     #: A unique identifier for the organization that made the call. This is a unique identifier across Cisco.
     org_uuid: Optional[str] = Field(alias='Org UUID')
     #: Populated for calls that transfer, hold, wait, and so on. For example:
@@ -173,22 +166,64 @@ class CDR(ApiModel):
     #: The time this report was created. Time is in UTC.
     report_time: Optional[datetime] = Field(alias='Report time')
     #: If present, this field's only reported in Originating records. Route group identifies the route group used for
-    # outbound calls routed via a route group to Premises-based PSTN or an on-prem deployment integrated with Webex
-    # Calling (dial plan or unknown extension).
+    #: outbound calls routed via a route group to Premises-based PSTN or an on-prem deployment integrated with Webex
+    #: Calling (dial plan or unknown extension).
     route_group: Optional[str] = Field(alias='Route group')
     #: The main number for the user's site where the call was made or received.
     site_main_number: Optional[str] = Field(alias='Site main number')
     #: Site timezone is the offset in minutes from UTC time of the user's timezone.
     site_timezone: Optional[str] = Field(alias='Site timezone')
-    #: This is the start time of the call, the answer time may be slightly after this. Time is in UTC.
-    start_time: Optional[datetime] = Field(alias='Start time')
     #: If the call is TO or FROM a mobile phone using Webex Go, the Client type will show SIP, and Sub client type
-    # will show MOBILE_NETWORK.
+    #: will show MOBILE_NETWORK.
     sub_client_type: Optional[str] = Field(alias='Sub client type')
-    #: The type of user (user or workspace) that made or received the call. For example:
-    user_type: Optional[Union[CDRUserType, str]] = Field(alias='User type')
     #: A unique identifier for the user associated with the call. This is a unique identifier across Cisco products.
     user_uuid: Optional[str] = Field(alias='User UUID')
+    #: The type of user (user or workspace) that made or received the call. For example:
+    user_type: Optional[Union[CDRUserType, str]] = Field(alias='User type')
+    #: For incoming calls, the telephone number of the user. For outgoing calls, it's the telephone number of the
+    #: called party.
+    called_number: Optional[str] = Field(alias='Called number')
+    #: For incoming calls, the telephone number of the calling party. For outgoing calls, it's the telephone number
+    #: of the user.
+    calling_number: Optional[str] = Field(alias='Calling number')
+    #: Location of the report.
+    location: Optional[str] = Field(alias='Location')
+    #: Dialed digits
+    #: The keypad digits as dialed by the user, before pre-translations. This field reports multiple call dial
+    #: possibilities:
+    #
+    #: Feature access codes (FAC) used for invoking features such as Last Number Redial or a Call Return.
+    #:
+    #: An extension that got dialed and a mis-dialed keypad digit from a device/app.
+    #:
+    #: When a user must dial an outside access code (for example, 9+) before dialing a number, this access code is
+    #: also reported, as well as the digits dialed thereafter. Note that when pre-translations have no effect,
+    #: the dialed digits field contains the same data as the called number field. This field is only used for
+    #: originating (outgoing) Calls and is not available for terminating (incoming) Calls.
+    dialed_digits: Optional[str] = Field(alias='Dialed digits')
+    #: Indicates which party released the call first. The possible values are:
+    #:
+    #: Local: Used when the local user has released the call first.
+    #: Remote: Used when the far end party releases the call first.
+    #: Unknown: Used when the call has partial information or is unable to gather enough information about the party
+    #:   who released the call. It could be because of situations like force lock or because of a session audit failure.
+    releasing_party: Optional[str] = Field(alias='Releasing party')
+    #: When the call has been redirected one or more times, this field reports the last redirecting number.
+    #: Identifies who last redirected the call. Only applies to call scenarios such as transfer, call forwarded calls,
+    #: simultaneous rings, etc.
+    redirecting_number: Optional[str] = Field(alias='Redirecting number')
+    #: A unique identifier for the site associated with the call. Unique across Cisco products.
+    site_uuid: Optional[str] = Field(alias='Site UUID')
+    #: A unique identifier for the user's department name.
+    department_id: Optional[str] = Field(alias='Department ID')
+    #: Transfer related call ID is used as a call identifier of the other call involved in the transfer. You can share
+    #: this ID with Cisco TAC to help them pinpoint parties who are involved during a call transfer.
+    transfer_related_call_id: Optional[str] = Field(alias='Transfer related call ID')
+    #: The authorization code admin created for a location or site for users to use. Collected by the
+    #: Account/Authorization Codes or Enhanced Outgoing Calling Plan services.
+    authorization_code: Optional[str] = Field(alias='Authorization code')
+    #: The device model type the user is using to make or receive the call.
+    model: Optional[str] = Field(alias='Model')
 
 
 @dataclass(init=False)